pyproject.toml
[project]
name = "splitmind"
version = "0.1.0"
description = "Privacy-preserving multi-agent task orchestration system"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "SplitMind Team"}
]
keywords = ["ai", "privacy", "multi-agent", "llm", "task-orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# 核心依赖 - 最小化
dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# AI 提供商支持
providers = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "httpx>=0.25.0",
]

# 命令行工具
cli = [
    "click>=8.0.0",
    "rich>=13.0.0",
]

# Web 界面
web = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
]

# 异步支持
async = [
    "aiohttp>=3.9.0",
    "pytest-asyncio>=0.23.0",
]

# 配置管理
config = [
    "pydantic-settings>=2.0.0",
]

# 完整安装
all = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "httpx>=0.25.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "aiohttp>=3.9.0",
    "jinja2>=3.1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.0.0",
]

# 开发工具
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
splitmind = "splitmind.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/splitmind"
Documentation = "https://github.com/yourusername/splitmind#readme"
Repository = "https://github.com/yourusername/splitmind"
Issues = "https://github.com/yourusername/splitmind/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["splitmind*"]
exclude = ["packages*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
